[not verified] Make button widths work
Make button widths work - they change the size of the buttons, but it's
done based upon --jetpack-payment-buttons-gap (0.5em) rather than the
actual block gap in use, so it isn't 100% correct yet.

Buttons widths are always in percents in a multipe of 25, so it can
hardcode a certain number of calculations. Payment buttons allow widths
of any number in percents, pixels and ems.